chore: Reorganizes Algorithms
