+++ packages/apollos-data-connector-rock/src/features/data-source.js
@@ -1,28 +1,10 @@
-import { flatten, get } from 'lodash';
+import { get } from 'lodash';
 import RockApolloDataSource from '@apollosproject/rock-apollo-data-source';
 import { createGlobalId } from '@apollosproject/server-core';
 import ApollosConfig from '@apollosproject/config';
 
 export default class Feature extends RockApolloDataSource {
   resource = '';
-
-  // Names of Action Algoritms mapping to the functions that create the actions.
-  ACTION_ALGORITHIMS = Object.entries({
-    // We need to make sure `this` refers to the class, not the `ACTION_ALGORITHIMS` object.
-    PERSONA_FEED: this.personaFeedAlgorithm,
-    CONTENT_CHANNEL: this.contentChannelAlgorithm,
-    SERMON_CHILDREN: this.sermonChildrenAlgorithm,
-    UPCOMING_EVENTS: this.upcomingEventsAlgorithm,
-    CAMPAIGN_ITEMS: this.campaignItemsAlgorithm,
-    SERIES_IN_PROGRESS: this.seriesInProgressAlgorithm,
-    USER_FEED: this.userFeedAlgorithm,
-    DAILY_PRAYER: this.dailyPrayerAlgorithm,
-  }).reduce((accum, [key, value]) => {
-    // convenciance code to make sure all methods are bound to the Features dataSource
-    // eslint-disable-next-line
-    accum[key] = value.bind(this);
-    return accum;
-  }, {});
 
   getFromId(args, id, { info } = { info: {} }) {
     this.cacheControl = info.cacheControl;
@@ -43,21 +25,6 @@ export default class Feature extends RockApolloDataSource {
     return JSON.stringify(args);
   }
 
-  async runAlgorithms({ algorithms }) {
-    // We should flatten just in case a single algorithm generates multiple actions
-    return flatten(
-      await Promise.all(
-        algorithms.map(async (algorithm) => {
-          // Lookup the algorithm function, based on the name, and run it.
-          if (typeof algorithm === 'object') {
-            return this.ACTION_ALGORITHIMS[algorithm.type](algorithm.arguments);
-          }
-          return this.ACTION_ALGORITHIMS[algorithm]();
-        })
-      )
-    );
-  }
-
   async createActionListFeature({
     algorithms = [],
     title,
@@ -65,7 +32,8 @@ export default class Feature extends RockApolloDataSource {
     primaryAction,
   }) {
     // Generate a list of actions.
-    const actions = () => this.runAlgorithms({ algorithms });
+    const { ActionAlgorithm } = this.context.dataSources;
+    const actions = () => ActionAlgorithm.runAlgorithms({ algorithms });
 
     // Ensures that we have a generated ID for the Primary Action related node, if not provided.
     if (
@@ -109,20 +77,23 @@ export default class Feature extends RockApolloDataSource {
     // Generate a list of actions.
     let actions;
     let heroCard;
+    const { ActionAlgorithm } = this.context.dataSources;
 
     // If we have a strategy for selecting the hero card.
     if (heroAlgorithms && heroAlgorithms.length) {
       // The actions come from the action algorithms
-      actions = () => this.runAlgorithms({ algorithms });
+      actions = () => ActionAlgorithm.runAlgorithms({ algorithms });
       // and the hero comes from the hero algorithms
       heroCard = async () => {
-        const cards = await this.runAlgorithms({ algorithms: heroAlgorithms });
+        const cards = await ActionAlgorithm.runAlgorithms({
+          algorithms: heroAlgorithms,
+        });
         return cards.length ? cards[0] : null;
       };
       // Otherwise, if we don't have a strategy
     } else {
       // Get all the cards (sorry, no lazy loading here)
-      const allActions = await this.runAlgorithms({ algorithms });
+      const allActions = await ActionAlgorithm.runAlgorithms({ algorithms });
       // The actions are all actions after the first
       actions = allActions.slice(1);
       // And the hero is the first action.
@@ -170,7 +141,8 @@ export default class Feature extends RockApolloDataSource {
     isFeatured = false,
   }) {
     // Generate a list of cards.
-    const cards = () => this.runAlgorithms({ algorithms });
+    const { ActionAlgorithm } = this.context.dataSources;
+    const cards = () => ActionAlgorithm.runAlgorithms({ algorithms });
     return {
       // The Feature ID is based on all of the action ids, added together.
       // This is naive, and could be improved.
@@ -199,7 +171,8 @@ export default class Feature extends RockApolloDataSource {
     primaryAction,
   }) {
     // Generate a list of horizontal cards.
-    const cards = () => this.runAlgorithms({ algorithms });
+    const { ActionAlgorithm } = this.context.dataSources;
+    const cards = () => ActionAlgorithm.runAlgorithms({ algorithms });
     // Ensures that we have a generated ID for the Primary Action related node, if not provided.
     if (
       primaryAction &&
@@ -253,7 +226,8 @@ export default class Feature extends RockApolloDataSource {
   }
 
   createPrayerListFeature({ algorithms = [], title, subtitle, isCard = true }) {
-    const prayers = () => this.runAlgorithms({ algorithms });
+    const { ActionAlgorithm } = this.context.dataSources;
+    const prayers = () => ActionAlgorithm.runAlgorithms({ algorithms });
     return {
       // The Feature ID is based on all of the action ids, added together.
       // This is naive, and could be improved.
@@ -272,186 +246,6 @@ export default class Feature extends RockApolloDataSource {
       // Typename is required so GQL knows specifically what Feature is being created
       __typename: 'PrayerListFeature',
     };
-  }
-
-  async dailyPrayerAlgorithm({ limit = 10 } = {}) {
-    this.setCacheHint({ maxAge: 0, scope: 'PRIVATE' });
-
-    const { PrayerRequest } = this.context.dataSources;
-    const cursor = await PrayerRequest.byDailyPrayerFeed();
-    return cursor.top(limit).get();
-  }
-
-  // Gets the first 3 upcoming events
-  async upcomingEventsAlgorithm() {
-    const { Event } = this.context.dataSources;
-
-    // Get the first three persona items.
-    const events = await Event.findRecent()
-      .top(3)
-      .get();
-    // Map them into specific actions.
-    return events.map((event, i) => ({
-      id: `${event.id}${i}`,
-      title: Event.getName(event),
-      subtitle: Event.getDateTime(event.schedule).start,
-      relatedNode: { ...event, __type: 'Event' },
-      image: Event.getImage(event),
-      action: 'READ_EVENT',
-      summary: '',
-    }));
-  }
-
-  // Gets the first 3 items for a user, based on their personas.
-  async personaFeedAlgorithm() {
-    this.setCacheHint({ maxAge: 0, scope: 'PRIVATE' });
-
-    const { ContentItem } = this.context.dataSources;
-
-    // Get the first three persona items.
-    const personaFeed = await ContentItem.byPersonaFeed(3);
-    const items = await personaFeed.expand('ContentChannel').get();
-
-    // Map them into specific actions.
-    return items.map((item, i) => ({
-      id: `${item.id}${i}`,
-      title: item.title,
-      subtitle: get(item, 'contentChannel.name'),
-      relatedNode: { ...item, __type: ContentItem.resolveType(item) },
-      image: ContentItem.getCoverImage(item),
-      action: 'READ_CONTENT',
-      summary: ContentItem.createSummary(item),
-    }));
-  }
-
-  // Gets a configurable amount of content items from a specific content channel.
-  async contentChannelAlgorithm({ contentChannelId, limit = null } = {}) {
-    if (contentChannelId == null) {
-      throw new Error(
-        `contentChannelId is a required argument for the CONTENT_CHANNEL ActionList algorithm.
-Make sure you structure your algorithm entry as \`{ type: 'CONTENT_CHANNEL', aruments: { contentChannelId: 13 } }\``
-      );
-    }
-
-    const { ContentItem } = this.context.dataSources;
-    const cursor = ContentItem.byContentChannelId(contentChannelId).expand(
-      'ContentChannel'
-    );
-
-    const items = limit ? await cursor.top(limit).get() : await cursor.get();
-
-    return items.map((item, i) => ({
-      id: `${item.id}${i}`,
-      title: item.title,
-      subtitle: get(item, 'contentChannel.name'),
-      relatedNode: { ...item, __type: ContentItem.resolveType(item) },
-      image: ContentItem.getCoverImage(item),
-      action: 'READ_CONTENT',
-      summary: ContentItem.createSummary(item),
-    }));
-  }
-
-  // Gets a configurable amount of content items that are a child of the most recent sermon.
-  async sermonChildrenAlgorithm({ limit = null } = {}) {
-    const { ContentItem } = this.context.dataSources;
-
-    const sermon = await ContentItem.getSermonFeed().first();
-    if (!sermon) {
-      return [];
-    }
-
-    const cursor = (await ContentItem.getCursorByParentContentItemId(
-      sermon.id
-    )).expand('ContentChannel');
-    const items = limit ? await cursor.top(limit).get() : await cursor.get();
-
-    return items.map((item, i) => ({
-      id: `${item.id}${i}`,
-      title: item.title,
-      subtitle: get(item, 'contentChannel.name'),
-      relatedNode: { ...item, __type: ContentItem.resolveType(item) },
-      image: ContentItem.getCoverImage(item),
-      action: 'READ_CONTENT',
-      summary: ContentItem.createSummary(item),
-    }));
-  }
-
-  // Gets a configurable amount of content items from each of the configured campaigns
-  async campaignItemsAlgorithm({ limit = 1 } = {}) {
-    const { ContentItem } = this.context.dataSources;
-
-    const channels = await ContentItem.byContentChannelIds(
-      ApollosConfig.ROCK_MAPPINGS.CAMPAIGN_CHANNEL_IDS
-    ).get();
-
-    const items = flatten(
-      await Promise.all(
-        channels.map(async ({ id, title }) => {
-          const childItemsCursor = await ContentItem.getCursorByParentContentItemId(
-            id
-          );
-
-          const childItems = await childItemsCursor
-            .top(limit)
-            .expand('ContentChannel')
-            .get();
-
-          return childItems.map((item) => ({
-            ...item,
-            channelSubtitle: title,
-          }));
-        })
-      )
-    );
-
-    return items.map((item, i) => ({
-      id: `${item.id}${i}`,
-      title: item.title,
-      subtitle: get(item, 'contentChannel.name'),
-      relatedNode: { ...item, __type: ContentItem.resolveType(item) },
-      image: ContentItem.getCoverImage(item),
-      action: 'READ_CONTENT',
-      summary: ContentItem.createSummary(item),
-    }));
-  }
-
-  async userFeedAlgorithm({ limit = 20 } = {}) {
-    const { ContentItem } = this.context.dataSources;
-
-    const items = await ContentItem.byUserFeed()
-      .top(limit)
-      .get();
-
-    return items.map((item, i) => ({
-      id: `${item.id}${i}`,
-      title: item.title,
-      subtitle: get(item, 'contentChannel.name'),
-      relatedNode: { ...item, __type: ContentItem.resolveType(item) },
-      image: ContentItem.getCoverImage(item),
-      action: 'READ_CONTENT',
-      summary: ContentItem.createSummary(item),
-    }));
-  }
-
-  async seriesInProgressAlgorithm({ limit = 3 } = {}) {
-    this.setCacheHint({ maxAge: 0, scope: 'PRIVATE' });
-
-    const { ContentItem } = this.context.dataSources;
-
-    const items = await (await ContentItem.getSeriesWithUserProgress())
-      .expand('ContentChannel')
-      .top(limit)
-      .get();
-
-    return items.map((item, i) => ({
-      id: `${item.id}${i}`,
-      title: item.title,
-      subtitle: get(item, 'contentChannel.name'),
-      relatedNode: { ...item, __type: ContentItem.resolveType(item) },
-      image: ContentItem.getCoverImage(item),
-      action: 'READ_CONTENT',
-      summary: ContentItem.createSummary(item),
-    }));
   }
 
   async getScriptureShareMessage(ref) {
